Remove LightningGradient
Interesting: I think Tailwind is purging classes that are only used in MDX files. When I removed this file, the lightning gradient stopped working until I moved it out of `@layer components`. Gotta figure out how to get it to not purge styles that are only used in MDX files